[refactor] home screen 코드 정리
