+++ lyc_flutter_project/lib/home/screen/home_screen.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:lyc_flutter_project/common/widget/custom_loading.dart';
 import 'package:lyc_flutter_project/common/widget/home_appbar.dart';
 import 'package:lyc_flutter_project/common/widget/preview_posting_card.dart';
 import 'package:lyc_flutter_project/data/app_color.dart';
 import 'package:lyc_flutter_project/common/widget/round_image.dart';
 import 'package:lyc_flutter_project/home/provider/home_provider.dart';
+import 'package:lyc_flutter_project/home/widget/director_card.dart';
+import 'package:lyc_flutter_project/home/widget/home_deco.dart';
+import 'package:lyc_flutter_project/home/widget/home_info.dart';
 import 'package:provider/provider.dart';
 
 class HomeScreen extends StatelessWidget {
@@ -40,23 +42,24 @@ class HomeScreen extends StatelessWidget {
                     SizedBox(
                       height: 150.0,
                       child: FutureBuilder(
-                        future: value.getPostingPreview(),
-                        builder: (context, snapshot) {
-                          if (value.loadingFeedPreview) return const Center(child: CustomLoading());
-                          if (value.feedPreviewList.isEmpty) return const Text("불러올 게시글이 없습니다");
-                          return ListView.builder(
-                            scrollDirection: Axis.horizontal,
-                            itemCount: value.feedPreviewList.length,
-                            itemBuilder: (context, index) {
-                              final posting = value.feedPreviewList[index];
-                              return PreviewPostingCard(
-                                postingId: posting.postingId,
-                                image: posting.image,
-                              );
-                            },
-                          );
-                        }
-                      ),
+                          future: value.getPostingPreview(),
+                          builder: (context, snapshot) {
+                            if (value.loadingFeedPreview)
+                              return const Center(child: CustomLoading());
+                            if (value.feedPreviewList.isEmpty)
+                              return const Text("불러올 게시글이 없습니다");
+                            return ListView.builder(
+                              scrollDirection: Axis.horizontal,
+                              itemCount: value.feedPreviewList.length,
+                              itemBuilder: (context, index) {
+                                final posting = value.feedPreviewList[index];
+                                return PreviewPostingCard(
+                                  postingId: posting.postingId,
+                                  image: posting.image,
+                                );
+                              },
+                            );
+                          }),
                     ),
                     // 오늘의 디렉터
                     const MarginBox(),
@@ -74,25 +77,9 @@ class HomeScreen extends StatelessWidget {
                               height: 90,
                               width: 90,
                               alignment: Alignment.center,
-                              child: Column(
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                  SizedBox(
-                                    height: 50,
-                                    width: 50,
-                                    child: RoundImage(
-                                      image: Image.asset(
-                                        'assets/ex_profile2.png',
-                                        fit: BoxFit.cover,
-                                      ),
-                                    ),
-                                  ),
-                                  const SizedBox(height: 3),
-                                  const Text(
-                                    'Winter',
-                                    style: TextStyle(color: AppColor.deepGrey),
-                                  ),
-                                ],
+                              child: const DirectorCard(
+                                image: "assets/ex_profile2.png",
+                                name: "winter",
                               ),
                             ),
                         ],
@@ -144,113 +131,7 @@ class HomeScreen extends StatelessWidget {
 
               // 유클리드
               const MarginBox(),
-              Container(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 30, vertical: 40),
-                color: const Color(0xffF4F5F6),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    const Text(
-                      '유클리드',
-                      style:
-                          TextStyle(fontWeight: FontWeight.bold, fontSize: 24),
-                    ),
-                    const MarginBox(),
-                    Row(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Expanded(
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              const DescriptTitle(label: 'Contact us'),
-                              const MiniBox(),
-                              const DescriptText(label: 'id@gmail.com'),
-                              const MiniBox(),
-                              const DescriptText(label: '+82 10-XXXX-XXXX'),
-                              const MiniBox(),
-                              const DescriptText(label: 'Address'),
-                              const SizedBox(height: 30),
-                              Row(
-                                children: [
-                                  SvgPicture.asset('assets/icon_facebook.svg'),
-                                  const SizedBox(width: 10),
-                                  SvgPicture.asset('assets/icon_linkedin.svg'),
-                                  const SizedBox(width: 10),
-                                  SvgPicture.asset('assets/icon_twitter.svg'),
-                                  const SizedBox(width: 10),
-                                  SvgPicture.asset('assets/icon_instagram.svg'),
-                                ],
-                              )
-                            ],
-                          ),
-                        ),
-                        const Expanded(
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              DescriptTitle(label: 'Products'),
-                              MiniBox(),
-                              DescriptText(label: 'Author volutpat.'),
-                              MiniBox(),
-                              DescriptText(label: 'Fermentum turpis.'),
-                              MiniBox(),
-                              DescriptText(label: 'Mi consequat.'),
-                              MiniBox(),
-                              DescriptText(label: 'Amet venenatis.'),
-                              MiniBox(),
-                              DescriptText(label: 'Convallis porttitor.'),
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                    const MarginBox(),
-                    Row(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        const Expanded(
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              DescriptTitle(label: 'About'),
-                              MiniBox(),
-                              DescriptText(label: 'Egestas vitae.'),
-                              MiniBox(),
-                              DescriptText(label: 'Viverra lorem ac.'),
-                              MiniBox(),
-                              DescriptText(label: 'Eget ac tellus.'),
-                              MiniBox(),
-                              DescriptText(label: 'Erat nulla.'),
-                              MiniBox(),
-                              DescriptText(label: 'Vulputate proin.'),
-                            ],
-                          ),
-                        ),
-                        Expanded(
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              const DescriptTitle(label: 'Get the app'),
-                              const MiniBox(),
-                              SvgPicture.asset('assets/icon_appstore.svg'),
-                              const MiniBox(),
-                              SvgPicture.asset('assets/icon_playstore.svg'),
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                    const MarginBox(),
-                    Text(
-                      'Copyright © 2024. All rights reserved.',
-                      style:
-                          TextStyle(color: AppColor.deepGrey.withOpacity(0.8)),
-                    )
-                  ],
-                ),
-              ),
+              const HomeInfo(),
             ],
           ),
         );
@@ -267,112 +148,3 @@ class HomeScreen extends StatelessWidget {
 }
 
 void _onTap() {}
-
-class MiniBox extends StatelessWidget {
-  const MiniBox({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox(height: 15);
-  }
-}
-
-class DescriptTitle extends StatelessWidget {
-  final String label;
-
-  const DescriptTitle({
-    super.key,
-    required this.label,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      label,
-      style: const TextStyle(
-        fontWeight: FontWeight.w600,
-        fontSize: 16,
-      ),
-    );
-  }
-}
-
-class DescriptText extends StatelessWidget {
-  final String label;
-
-  const DescriptText({super.key, required this.label});
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      label,
-      style: const TextStyle(
-        fontSize: 14,
-        fontWeight: FontWeight.w400,
-        color: Color(0xff353945),
-      ),
-    );
-  }
-}
-
-class Line extends StatelessWidget {
-  const Line({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return const Divider(
-      color: Color(0xffE4E4E4),
-      thickness: 2,
-    );
-  }
-}
-
-class TitleRow extends StatelessWidget {
-  final String label;
-  final VoidCallback onTap;
-
-  const TitleRow({
-    super.key,
-    required this.label,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(
-          label,
-          style: const TextStyle(
-            color: Colors.black,
-            fontWeight: FontWeight.bold,
-            fontSize: 18,
-          ),
-        ),
-        GestureDetector(
-          onTap: onTap,
-          child: SvgPicture.asset(
-            'assets/icon_plus.svg',
-            color: const Color(0xffA6A6A6),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class MarginBox extends StatelessWidget {
-  const MarginBox({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox(height: 60);
-  }
-}
